backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aiosmtplib==1.1.6
alembic==1.12.0
algoliasearch==3.0.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.3.0
autoflake==2.2.1
bcrypt==4.0.1
black==23.9.1
blurhash-python==1.2.1
boto3==1.28.54
botocore==1.31.54
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
coverage==7.4.0
croniter==1.4.1
cryptography==41.0.4
decorator==4.4.2
Deprecated==1.2.13
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
Faker==19.10.0
fastapi==0.103.1
ffmpeg-python==0.2.0
filelock==3.12.4
flake8==6.1.0
frozenlist==1.4.0
future==0.18.3
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
identify==2.5.29
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.0.0b2
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
orjson==3.9.7
packaging==23.1
pathspec==0.11.2
phonenumbers==8.12.48
Pillow==10.1.0
pkce==1.0.3
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.4.0
proglog==0.1.10
protobuf==4.24.3
psycopg2==2.9.7
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.10.4
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-env==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.2
segno==1.6.0
sentry-sdk==1.31.0
six==1.16.0
slack-sdk==3.22.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
temporalio==1.3.0
tldextract==3.1.0
tomli==2.0.1
tqdm==4.66.1
twilio==7.9.1
types-protobuf==4.24.0.2
typing_extensions==4.7.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
